requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
python-chess>=1.10.0
cairosvg>=2.7.0
Pillow>=10.0.0
imageio[pyav]>=2.31.0
stockfish>=3.28.0
